# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/integration_connectors:
    get:
      tags:
        - IntegrationConnectors
      summary: List IntegrationConnectors
      description: Get a paginated list of IntegrationConnectors with optional filtering and sorting
      operationId: listIntegrationConnectors
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationConnectorList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - IntegrationConnectors
      summary: Create IntegrationConnector
      description: Create a new IntegrationConnector
      operationId: createIntegrationConnector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntegrationConnectorRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationConnector'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/integration_events:
    get:
      tags:
        - IntegrationEvents
      summary: List IntegrationEvents
      description: Get a paginated list of IntegrationEvents with optional filtering and sorting
      operationId: listIntegrationEvents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationEventList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - IntegrationEvents
      summary: Create IntegrationEvent
      description: Create a new IntegrationEvent
      operationId: createIntegrationEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntegrationEventRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    IntegrationConnector:
      type: object
      required:
        - id
        - company_id
        - provider
        - kind
        - direction
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        provider:
          type: string
        kind:
          type: string
        direction:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateIntegrationConnectorRequest:
      type: object
      required:
        - company_id
        - provider
        - kind
        - direction
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        provider:
          type: string
        kind:
          type: string
        direction:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    IntegrationConnectorList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationConnector'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    IntegrationEvent:
      type: object
      required:
        - id
        - company_id
        - connector_id
        - event_type
        - external_id
        - business_key
        - status
        - payload
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        connector_id:
          type: string
          format: uuid
        event_type:
          type: string
        external_id:
          type: string
        business_key:
          type: string
        status:
          type: string
        payload:
          type: string
        mapped_ref_type:
          nullable: true
          type: string
        mapped_ref_id:
          nullable: true
          type: string
          format: uuid
        error_detail:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateIntegrationEventRequest:
      type: object
      required:
        - company_id
        - connector_id
        - event_type
        - external_id
        - business_key
        - status
        - payload
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        connector_id:
          type: string
          format: uuid
        event_type:
          type: string
        external_id:
          type: string
        business_key:
          type: string
        status:
          type: string
        payload:
          type: string
        mapped_ref_type:
          nullable: true
          type: string
        mapped_ref_id:
          nullable: true
          type: string
          format: uuid
        error_detail:
          nullable: true
          type: string
        metadata:
          type: object
...
    IntegrationEventList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationEvent'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []